refactor: extract MCP schema helpers
